Rename getCauses and export as getErrors
